+++ rollup.config.js
@@ -3,6 +3,7 @@ import buble from 'rollup-plugin-buble'
 export default {
   input: 'index.es.js',
   output: {
+    exports: 'default',
     file: 'index.cjs.js',
     format: 'cjs'
   },