+++ test/postcss/rollup.config.js
@@ -1,0 +1,17 @@
+import scss from '../../index.es.js'
+import postcss from 'postcss'
+import autoprefixer from 'autoprefixer'
+
+export default {
+  input: '../input.js',
+  output: {
+    file: 'output.js',
+    format: 'esm',
+  },
+  plugins: [
+    scss({
+      processor: css =>
+        postcss([autoprefixer({ overrideBrowserslist: 'Edge 18' })]),
+    }),
+  ],
+}